separate expression converters from Parser module
