+++ src/parse/Parser/_Parser.js
@@ -1,5 +1,4 @@
 import { create, hasOwn } from 'utils/object';
-import getConditional from './expressions/conditional';
 
 var Parser, ParseError, leadingWhitespace = /^\s+/;
 
@@ -61,16 +60,6 @@ Parser.prototype = {
 		}
 
 		return null;
-	},
-
-	readExpression: function () {
-		// The conditional operator is the lowest precedence operator (except yield,
-		// assignment operators, and commas, none of which are supported), so we
-		// start there. If it doesn't match, it 'falls through' to progressively
-		// higher precedence operators, until it eventually matches (or fails to
-		// match) a 'primary' - a literal or a reference. This way, the abstract syntax
-		// tree has everything in its proper place, i.e. 2 + 3 * 4 === 14, not 20.
-		return getConditional( this );
 	},
 
 	getLinePos: function ( char ) {
